Stress testing groundwork
This is some pre-gorundwork for stress testing

1. Make the console UI a full fledged client so lots of folks can
   use it
2. Refactor tests accordingl
3. Add a make-stress client which starts building my stress test
   cases, but its pretty bad and lots of things seem quadratic in
   very large zone count.

Address #1282
